backend/requirements.txt
falcon
gunicorn
waitress
psycopg2
falcon_cors
PyJwt 
werkzeug
bcrypt